get `pytest .` running with a failing test. rename test_actions.py -> prapti_test_actions.py for compatibility with pytest file name matching.
